homework6/typicode/pyproject.toml
[tool.poetry]
name = "typicode"
version = "0.1.0"
description = ""
authors = ["Shaj <andrey-shaj@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^2.7.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
